Section3/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
charset-normalizer==2.1.1
click==8.1.3
DateTime==4.5
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
xgboost==0.90
xmltodict==0.13.0
zipp==3.8.1
zope.interface==5.4.0
